requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-74hc595==1.4.6
adafruit-circuitpython-ads1x15==2.4.2
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-charlcd==3.5.1
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-dht==4.0.7
adafruit-circuitpython-mcp230xx==2.5.16
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
bluedot==2.0.0
click==8.1.8
dbus-python==1.4.0
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pigpio==1.78
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
python-etcd==0.4.5
pyusb==1.3.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3